Include Patch 13111
Related to the discussion here: http://avalondock.codeplex.com/discussions/400843

This adds a property to the LayoutGroup class named 'CacheTabsContent', defaults to false but when set to true, it uses the class TabContent from: http://www.codeproject.com/Articles/460989/WPF-TabControl-Turning-Off-Tab-Virtualization to cache to content of the tabcontrol that holds the LayoutDocumentControl/LayoutAnchorableControl.

I didnt test much yet, but would like to have feedback if a solution like that would be OK to be integrated into AvalonDock.

This patch modifies all existing themes to integrate the functionality.